reorganize loginform components, minor bugfix
